+++ ChangeLog.md
@@ -9,6 +9,7 @@
 
 * Be more resilient about missing metadata in CycloneDX SBOMs.
 * `project createbom` uses purl from SW360 if available instead of building it
+* `project createbom` adds SW360 project name, version and description to SBOM.
 
 ## 2.0.0 (2023-06-02)